requirement.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.0.2
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2026.1.post1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.39.0